Change ahab.bb for new fsimage.sh Version

+++ meta-fus-bsp/recipes-security/ahab/ahab.bb
@@ -16,17 +16,30 @@ B = "${WORKDIR}/build"
 PACKAGE_ARCH = "${MACHINE_ARCH}"
 COMPATIBLE_MACHINE = "(mx8|mx93)"
 
+### default values
 SRK_index ?= "0"
 SRK_filename ?= "SRK1_sha384_secp384r1_v3_usr_crt.pem"
 SRK_revoke ?= "0x0"
-### skript datei
+
+### NBoot Binary
+nboot_file ?= "nboot.fs"
+
+### script files
 Update_Files ?= ""
 Update_Scripts ?= ""
 
+### sign linux if fitimage is there
 SIGN_LINUX = "${@bb.utils.contains('KERNEL_CLASSES', 'kernel-fitimage', 'true', 'false', d)}"
 
-DEPENDS:append = " imx-cst-native nboot u-boot-fus xxd-native u-boot-tools-native linux-fus optee-os imx-atf"
+DEPENDS:append = " imx-cst-native u-boot-fus xxd-native u-boot-tools-native linux-fus optee-os imx-atf"
 inherit deploy
+
+### recipe should alway run, do not stamp the tasks
+python __anonymous () {
+    tasks = d.keys()
+    for t in tasks:
+        d.setVarFlag(t, "nostamp", "1")
+}
 
 do_configure[depends] += " linux-fus:do_deploy "
 do_configure() {
@@ -51,7 +64,11 @@ do_configure() {
 	sed -i "s/###index###/${SRK_index}/g" ${B}/input_edited.csf
 	sed -i "s/###revoke###/${SRK_revoke}/g" ${B}/input_edited.csf
 
-	cp ${DEPLOY_DIR_IMAGE}/Firmware/nboot.fs ${B}/nboot_signed.fs
+	if [ -f ${DL_DIR}/${nboot_file} ]; then
+		cp ${DL_DIR}/${nboot_file} ${B}/nboot_signed.fs
+	elif [ -f ${DEPLOY_DIR_IMAGE}/Firmware/${nboot_file} ]; then
+		cp ${DEPLOY_DIR_IMAGE}/Firmware/${nboot_file} ${B}/nboot_signed.fs
+	fi
 	cp ${DEPLOY_DIR_IMAGE}/Firmware/uboot-${MACHINE_ARCH}.fs ${B}/uboot-${MACHINE_ARCH}_signed.fs
 
 	if ${SIGN_LINUX}; then
@@ -69,21 +86,38 @@ do_configure() {
 }
 
 do_compile() {
-	for file in nboot uboot-${MACHINE_ARCH}
+	if [ -f ${B}/nboot_signed.fs ]; then
+		image_list="nboot uboot-${MACHINE_ARCH}"
+	else
+		image_list="uboot-${MACHINE_ARCH}"
+	fi
+
+	for file in ${image_list}
 	do
-		cat ${DEPLOY_DIR_IMAGE}/Firmware/${file}.fs | ${WORKDIR}/fsimage.sh | grep "i.MX Container: type: OEM" | while read line
+		### we need container and signature offsets, we parse the first from the fsimage.sh output and get the second from the container header
+		cat ${B}/${file}_signed.fs | ${WORKDIR}/fsimage.sh | grep "IMX Container Header" | grep -v "NXP signed" | while read line
 		do
-			container=$(echo $line | sed 's/.*Container offset: //' | sed 's/,.*//')
-			signature=$(echo $line | sed 's/.*Signature block offset: //')
+			container_hex=0x$(echo $line | sed 's/.*: 0*//' | sed 's/ .*//')
+			container_dec=$(printf %d $container_hex)
+			signature_rel_hex=0x$(xxd -e -l 2 -s $(expr $container_dec + 12) ${B}/${file}_signed.fs  | awk '{print $2}'); 
+			signature_rel_dec=$(printf %d $signature_rel_hex);
+			signature_dec=$(expr $container_dec + $signature_rel_dec);
+			signature_hex=0x$(printf %x $signature_dec);
+
 			cp ${B}/input_edited.csf ${B}/input_edited2.csf
-			sed -i "s/###header###/0x${container}/g" ${B}/input_edited2.csf
-			sed -i "s/###signature###/0x${signature}/g" ${B}/input_edited2.csf
+			sed -i "s/###header###/${container_hex}/g" ${B}/input_edited2.csf
+			sed -i "s/###signature###/${signature_hex}/g" ${B}/input_edited2.csf
 			sed -i "s/###filename###/${file}_signed.fs/g" ${B}/input_edited2.csf
+
+			bbnote $(cat  ${B}/input_edited2.csf)
+
 			cst -i ${B}/input_edited2.csf -o ${B}/${file}_signed.fs
 			echo ${file}_signed.fs is signed
 		done
 	done
-	cat ${B}/nboot_signed.fs ${B}/uboot-${MACHINE_ARCH}_signed.fs > flash_signed.fs
+	if [ -f ${B}/nboot_signed.fs ]; then
+		cat ${B}/nboot_signed.fs ${B}/uboot-${MACHINE_ARCH}_signed.fs > flash_signed.fs
+	fi
 
 	if ${SIGN_LINUX}; then
 		sed -i "s/###fitimage###/fitImage-${MACHINE_ARCH}.bin/g" ${B}/os_cntr_edited.cfg
@@ -104,7 +138,6 @@ do_compile() {
 		cst -i ${B}/input_edited2.csf -o ${B}/os_cntr_signed.cntr
 	fi
 
-	### jetzt noch das update script
 	for i in ${Update_Files}; do
 		cp ${WORKDIR}/os_cntr.cfg ${B}/script.cfg #alles von W zu B
 		sed -i "s/###fitimage###/${i}.scr/g" ${B}/script.cfg
@@ -137,13 +170,14 @@ addtask deploy after do_compile
 do_deploy() {
 	install -d ${DEPLOY_DIR_IMAGE}/Secure
 	install -m 0644 ${B}/uboot-${MACHINE_ARCH}_signed.fs ${DEPLOY_DIR_IMAGE}/Secure
-	install -m 0644 ${B}/nboot_signed.fs ${DEPLOY_DIR_IMAGE}/Secure
-	install -m 0644 ${B}/flash_signed.fs ${DEPLOY_DIR_IMAGE}/Secure
+	if [ -f ${B}/nboot_signed.fs ]; then
+		install -m 0644 ${B}/nboot_signed.fs ${DEPLOY_DIR_IMAGE}/Secure
+		install -m 0644 ${B}/flash_signed.fs ${DEPLOY_DIR_IMAGE}/Secure
+	fi
 	if ${SIGN_LINUX}; then
 		install -m 0644 ${B}/os_cntr_signed.cntr ${DEPLOY_DIR_IMAGE}/Secure
 	fi
 
-	### update skript
 	for i in ${Update_Files}; do
 		name=$(echo ${i} | sed "s/\..*//g")_signed.scr
 		cp ${B}/${i}.scr.cntr.signed ${DEPLOY_DIR_IMAGE}/Secure/${name}
